backend/requirements.txt
fastapi
uvicorn[standard]
weaviate-client>=4.18.3
sentence-transformers
python-multipart
pydantic
requests
gradio
pdfplumber
python-dotenv
openai
pytesseract
pillow
python-dotenv